requirements.txt
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
chromedriver==2.24.1
chromedriver-binary==79.0.3945.16.0
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
isort==4.3.20
jedi==0.15.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.1.1
mccabe==0.6.1
mpmath==1.1.0
netifaces==0.10.9
numpy==1.17.4
parso==0.5.1
psutil==5.6.5
psycopg2==2.8.4
pylint==2.3.1
pyobjc-core==6.1
pyobjc-framework-Cocoa==6.1
pyparsing==2.4.5
PySide2==5.13.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rumps==0.3.0
selenium==3.141.0
shiboken2==5.13.0
six==1.13.0
sqlparse==0.3.0
sympy==1.4
typed-ast==1.4.0
urllib3==1.25.7
whitenoise==4.1.4
wrapt==1.11.2
youtube-dl==2019.11.22
